Rename `types` to `type_defs`
